start coding list

+++ lib/my_help/modify.rb
@@ -11,14 +11,23 @@ module MyHelp
     end
 
     def delete(help_file)
-      FileUtils.rm(help_file, verbose: @conf[:verbose])
+      if File.exist?(help_file)
+        FileUtils.rm(help_file, verbose: @conf[:verbose])
+      else
+        puts "file #{help_file} does not exist."
+      end
     end
 
     def edit(help_name)
       p help_file = File.join(@conf[:local_help_dir],
                               help_name + @conf[:ext])
-      p comm = "#{@conf[:editor]} #{help_file}"
-      system(comm)
+      if File.exist?(help_file)
+        p comm = "#{@conf[:editor]} #{help_file}"
+        system(comm)
+      else
+        puts "file #{help_file} does not exist,"
+        puts "make #{help_name} first by 'new' command."
+      end
     end
   end
 end
